backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
python-dateutil==2.8.2
Werkzeug==2.3.7
bcrypt==4.1.2

PyJWT==2.8.0